requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
plotly==5.18.0
pydantic==1.10.13
langchain==0.1.0
langchain-groq==0.0.1
langchain-core==0.1.23
langchain-community==0.0.20
langsmith==0.0.87